+++ NEWS.md
@@ -1,3 +1,8 @@
 # cachefun 0.1.0
 
-* Added a `NEWS.md` file to track changes to the package.
+* Initial version
+
+# cachefun 0.1.1
+
+* Variable default settings
+* Reset of internal cache